latest version
separated the fact forecasting medewerker into 3 separate fact tables. one for salaris, one for softlanding, and the other for the rest of the looncomponents. the reason being that the first construction of the fact forecasting medewerker made it so that filter propagation was not optimal, and therefore gave duplicate or not properly filtered values. by separating the fact tables, the salaries could be calculated properly and used for othe looncomponents as calculated columns in the other fact table